swapping create_training_data files
